keelbug
Working keelbug with spawnable biped and actors with AI
AI is modeled off of flood infection forms

Needs movement animations, currently only uses the idle animations that I found data for in the Digsite release

Packaged and exported the model myself, shaders are a bit janky on this version but its fine for now. Collision on the model is currently weird. May produce errors.

Currently they can spawn in and they will jump at the player and damage them.