engineer
This is mostly just the current engineer found in the Digsite release but I've modified some of the tags in attempts to get them to use weapons/throw grenades, etc. and have death animations

Currently you can spawn them in and if you make sure all their start point and movement nodes are midair, they will fly around and come up to the player. They are fun to shoot but T-pose on death currently with chance to drop plasma grenades